moved restore endpoint into separate controller
