+++ backend/Tasque.WebAPI/Controllers/AuthController.cs
@@ -1,12 +1,7 @@
-﻿using Mailjet.Client;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Tasque.Core.BLL.JWT;
 using Tasque.Core.BLL.Services.Auth;
-using Tasque.Core.BLL.Services.Email;
 using Tasque.Core.Common.DTO;
-using Tasque.Core.Common.Models.Email;
 
 namespace Tasque.Core.WebAPI.Controllers
 {
@@ -15,15 +10,11 @@ namespace Tasque.Core.WebAPI.Controllers
     [ApiController]
     public class AuthController : ControllerBase
     {
-        private AuthService _service;
-        private PasswordResetService _passwordService;
-        private IEmailService _emailService;
+        private AuthService _service;        
 
-        public AuthController(AuthService service, PasswordResetService passwordService, IEmailService emailService)
+        public AuthController(AuthService service)
         {
             _service = service;
-            _passwordService = passwordService;
-            _emailService = emailService;
         }
 
         [HttpPost("login")]
@@ -51,27 +42,6 @@ namespace Tasque.Core.WebAPI.Controllers
         {
             await _service.Register(registerInfo);
             return Ok();
-        }
-
-        [HttpPost("restore/request")]
-        public async Task<IActionResult> RequestPasswordRestore([FromQuery] string email)
-        {
-            await _passwordService.Request(email);
-            return Ok();
-        }
-
-        [HttpGet("restore")]
-        public async Task<IActionResult> ValidatePasswordRestoreToken([FromQuery] Guid token)
-        {
-            await _passwordService.ValidateToken(token);
-            return Ok();
-        }
-
-        [HttpPost("restore")]
-        public async Task<IActionResult> RequestPasswordRestore(PasswordChangeDto body)
-        {
-            string token = await _passwordService.Confirm(body);
-            return Ok(token);
         }
     }
 }
